==> 03 Scripts/1. Data Wrangling.ipynb <==
# 01. Import libraries

# Import libraries

import pandas as pd
import numpy as np
import os

# 02. Setup project path

# Create path to project folder

projpath = r'/Users/laineyodette/Documents/proDocs/professional development/Learning/Learning - Data Analytics/CareerFoundry - Become a Data Analyst/Data Immersion Course/A4 - Python Fundamentals for Data Analysts/Submitted Tasks/2024-01 Instacart Basket Analysis'

# 03. Create dataframes

# Import orders.csv file to create dataframe 

df_ords = pd.read_csv(os.path.join(projpath, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# Import products.csv file to create dataframe 

df_prods = pd.read_csv(os.path.join(projpath, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Import departments.csv file to create dataframe 

df_dep = pd.read_csv(os.path.join(projpath, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

# 04. Wrangle the data

# Remove the 'eval_set' column from the df_ords dataframe

df_ords = df_ords.drop(columns = ['eval_set'])

# Count values are in the 'days since prior order' column - including missing NaN values

df_ords['days_since_prior_order'].value_counts(dropna = False)

# Rename the 'order_dow' column to be more intuitive

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

# Change the data type of the order_ID column to string and override as new type

df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

df_dep.head()

# Transpose df_dep

df_dep.T

# Transpose df_dep and create new df

df_dep_t = df_dep.T

df_dep_t

# Add an index

df_dep_t.reset_index()

# Create a new header

new_header = df_dep_t.iloc[0]

new_header

# Create a new dataframe that only copies over rows beyond the first row

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# Assign new_header names as the column names in the new df

df_dep_t_new.columns = new_header

df_dep_t_new

# 05. Create a Data Dictionary

# Change departments dataframe into a data dictionary

data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

# Lookup what department is dept_id is 19

print(data_dict.get('19'))

# 06. Subsetting

# Create a subset for only the snacks department

df_snacks = df_prods[df_prods['department_id']==19]

df_prods['department_id']==19

df_prods[df_prods['department_id']==19]

df_snacks.head()

# 07. Task Steps

# Step 2 - Find another unnecessary identifier variable in the df_ords dataframe and change it to a suitable format.

# Change the data type of the user_ID column to string and override as new type

df_ords['user_id'] = df_ords['user_id'].astype('str')

# Check Step 2 work

df_ords['user_id'].dtype

# Step 3 - Look for a variable in your df_ords dataframe and change its name without overwriting the dataframe.

# Rename the 'order_number' column to be more intuitive

df_ords.rename(columns = {'order_number': 'user_order_number'}, inplace = False)

# Step 4 - Find the frequency of the variable for the "busiest hour for orders" and share your findings.

# Count of values in the 'days since prior order' column 

df_ords['order_hour_of_day'].value_counts()

# Answer to Step 4 question "what is the busiest hour for placing orders": 10:00am

# Step 5 - Determine the meaning of 4 in the "department_id" column in the df_prods dataframe using data dictionary.

# Lookup what department has a department_id of 4

print(data_dict.get('4'))

# Answer to question in Step 5: the department whose ID is 4 is the produce department.

# Step 6 - Create a subset containing only information on breakfast item sales.

# Create a subset for only the breakfast item department (14)

df_breakfast = df_prods[df_prods['department_id']==14]

# Check work from Step 6 by displaying basic info on new df_breakfast dataframe

df_breakfast

# Step 7 - Find all observations about items from the alcohol, deli, beverages, and meat/seafood departments.

# Create a subset for only the dinner party items from various departments (alcohol 5, deli 20, beverages 7, and meat/seafood 12)

df_party = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

# Check work from Step 7 by displaying basic info on new df_party dataframe

df_party

# Step 8 - How many rows does the last dataframe you created have? Per the "check work" step I performed, 7650 rows, but just for fun, let's run some code to verify

# Check number of rows (and columns) of df_party dataframe

df_party.shape

# I have have confirmed the question asked in Step 8 of how many rows the df_party dataframe has: 7650

# Step 9 - Extract everything you can about customer of 'user_id' = '1'

# Create a new df for customer with user_id of '1'

df_user1 = df_ords[df_ords['user_id']=='1']

# Look at the new subset dataframe created from above to answer question posed in Step 9 to extract everything about user_id=1

df_user1

# continue looking at user_id = '1' for Step 9 question

df_user1.info()

# Step 10 - Provide some basic stats about the behavior of customer of 'user_id' = '1'

df_user1.describe()

# 08. Exporting wrangled dataframes

# Step 12 - Export the wrangled orders dataframe to 'Prepared Data' folder.

# Export wrangled orders dataframe 

df_ords.to_csv(os.path.join(projpath, '02 Data','Prepared Data', 'orders_wrangled.csv'))

# Step 13 - Export the wrangled departments dataframe to 'Prepared Data' folder.

# Export wrangled department dataframe 

df_dep_t_new.to_csv(os.path.join(projpath, '02 Data','Prepared Data', 'departments_wrangled.csv'))
