==> 03 Scripts/2. Data Consistency Checks.ipynb <==
# 01. Importing Libraries and Other Setup

# Import libraries

import pandas as pd
import numpy as np
import os

# Create quick path to project folder

projpath = r'/Users/laineyodette/Documents/proDocs/professional development/Learning/Learning - Data Analytics/CareerFoundry - Become a Data Analyst/Data Immersion Course/A4 - Python Fundamentals for Data Analysts/Submitted Tasks/2024-01 Instacart Basket Analysis'

# Import products.csv dataset into df_prods df

df_prods = pd.read_csv(os.path.join(projpath, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Import products.csv dataset into df_ords df

df_ords = pd.read_csv(os.path.join(projpath, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = 0)

df_ords.head()

# 02. Data Consistency Checks Practice

# Create a test dataframe

df_test = pd.DataFrame()

# Create a mixed type column

df_test['mix'] = ['a','b', 1, True]

df_test.head()

# Check for mixed types

for col in df_test.columns.tolist():
  weird = (df_test[[col]].map(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Convert mix column to string data type

df_test['mix'] = df_test['mix'].astype('str')

# 03. Data Consistency Checks - df_prods

# Conduct initial descriptive analysis on prods df

df_prods.describe()

# Find missing values by summing the number of null values in each column

df_prods.isnull().sum()

# Create a new df to view null values

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# Check to see how many rows are in the original dataframe

df_prods.shape

# Create a new dataframe without the rows with missing values

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

# Look for full duplicates in products df

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

# Create new df removing duplicates 

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# Export the checked products dataframe to 'Prepared Data' folder

df_prods_clean_no_dups.to_csv(os.path.join(projpath, '02 Data','Prepared Data', 'products_checked.csv'))

# 04. TASK Data Consistency Checks - df_ords

## Step 2 - Descriptive Analysis

# Conduct initial descriptive analysis on orders df

df_ords.describe()

**Observations**

- *Day of the week* is all under 7 (good since only 7 days in week and values should be between 0-6) - no outliers for min or max. 50th percentile is 3, which is correct.

- *Hour of the day* is all under 24 (good since only 24 hours in day and values should be between 0-23)

- *Days since prior order* should all be positive values, so these checks look okay.

## Step 3 - Mixed Data Types

# Check for mixed types

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

**Observations and Response**

- No mixed data type columns found in df_ords dataframe

- No fix needed (Step 4)

## Step 5 - Missing Values

# Find missing values by summing the number of null values in each column

df_ords.isnull().sum()

**Observations and Response**

- Missing values found in the "Days Since Prior Order" column

- This can be logical if a customer has not placed an order prior to the one in that row - in other words, this column would be null each time a new customer places their first order

## Step 6 - Fixing Missing Values

# Create boolean column for when the column days_since_prior_order is null to indicate it is the customer's first order

df_ords['customers_first_order'] = df_ords['days_since_prior_order'].isnull()*1

df_ords

**Action Taken**

As I suspected that the 206,209 missing values corresponded to the first order of each customer, I looked at the descriptive stats above in the userID column. The maximum value was indeed 206209, meaning there are 206,209 customers, so my suspicions were true.

I decided to create a new column (a flag column) to indicate that the row with null values for "Days Since Prior Order" is the customer's first order (or not) - it indicates 1 or "true" if it's the customer's first order. It may be something interesting to call out in the report later to see how many subsequent orders were placed vs how many customers only placed a first order.

Note that I did not choose to impute, as the values missing are the story. There are no actual missing values. This is the same reason I did not remove the data - as the fact that the data is missing is the story itself.

## Step 7 - Duplicate Values

# Look for full duplicates in orders df

df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

**Observations and Reactions**

- No duplicates were found in the orders df

- No action taken to address duplicates for Step 8

## Step 9 - Export Final Cleaned Data File

# Export the checked orders dataframe to 'Prepared Data' folder

df_ords.to_csv(os.path.join(projpath, '02 Data','Prepared Data', 'orders_checked.csv'))

**Note: df_prods file exported in Section 03 of code**
