==> 03 Scripts/3. Combining and Exporting Data.ipynb <==
# 01. Notebook Setup

# Import libraries

import pandas as pd
import numpy as np
import os

# Create quick path to project folder

projpath = r'/Users/laineyodette/Documents/proDocs/professional development/Learning/Learning - Data Analytics/CareerFoundry - Become a Data Analyst/Data Immersion Course/A4 - Python Fundamentals for Data Analysts/Submitted Tasks/2024-01 Instacart Basket Analysis'

# 02. File Import and Check

# Import orders_products_combined.csv data file into new df

df_ords_prods = pd.read_pickle(os.path.join(projpath, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# Check dimensions of new df - s/b 32,434,489 rows

df_ords_prods.shape

# Check column headers of df

df_ords_prods.head()

#Drop _merge columns in df_ords_prods df to cleanup

df_ords_prods = df_ords_prods.drop(columns = ['_merge'])

df_ords_prods.head()

# Import products_checked.csv data file into new df

df_prods = pd.read_csv(os.path.join(projpath, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = 0)

# Check dimensions of new df

df_prods.shape

# Check column headers of df

df_prods.head()

# 03. Merge and Check for Full Match

- Note common column headers between two dfs from head() code results above
- Use 'product_id' for merge key
- Use an Inner Join to ensure dataframe is complete without missing information - will include only information that has data from both frames in the merge

# Test merge

pd.merge(df_ords_prods, df_prods, on = 'product_id')

# Merge df_ords_prods with df_prods using inner join merge and merge flag

df_ords_prods_merged = df_ords_prods.merge(df_prods, on = 'product_id', indicator = True)

df_ords_prods_merged.head()

# Check the merge flag to ensure all rows from both dfs were included

df_ords_prods_merged['_merge'].value_counts()

**Confirmed** Merge only includes data that was in both dfs - count of rows is different than those of the individual dfs

# 04 Export merged df as pickle file

#Export new merged df as pickle to Prepared Data folder

df_ords_prods_merged.to_pickle(os.path.join(projpath, '02 Data', 'Prepared Data', 'ords_prods_merged.pkl'))
