==> 03 Scripts/4. Deriving New Variables.ipynb <==
# 01. Notebook Setup

# Import libraries

import pandas as pd
import numpy as np
import os

# Create quick path to project folder

projpath = r'/Users/laineyodette/Documents/proDocs/professional development/Learning/Learning - Data Analytics/CareerFoundry - Become a Data Analyst/Data Immersion Course/A4 - Python Fundamentals for Data Analysts/Submitted Tasks/2024-01 Instacart Basket Analysis'

# 02. File Import and Check

# Import orders_products_prior.csv data file into new df

df_ords_prods_all = pd.read_pickle(os.path.join(projpath, '02 Data', 'Prepared Data', 'ords_prods_merged.pkl'))

# Check dimensions of new df - should have 32,404,859 rows

df_ords_prods_all.shape

**Confirmed** Number of rows is as expected.

df_ords_prods_all.head()

# 03. Create subset to reduce amount of data

# Create subset of only first 1,000,000 rows

df = df_ords_prods_all[:1000000]

df.shape

# 04. Define a 'price range' function with if-else and apply to subset

# Define the price range function

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply price range function to subset and create new column for the price range results

df['price_range'] = df.apply(price_label, axis=1)

# Check the value counts on the price range column

df['price_range'].value_counts()

**Note** No high-range products found. Reminder: this is a subset.

# Check to see what the most expensive item in the subset is

df['prices'].max()

# 05. Define a 'price range' function with loc() and apply to subset

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts()

# 06. Define a 'price range' function with loc() and apply to whole df

df_ords_prods_all.loc[df_ords_prods_all['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_all.loc[(df_ords_prods_all['prices'] <= 15) & (df_ords_prods_all['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df_ords_prods_all.loc[df_ords_prods_all['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_all['price_range_loc'].value_counts()

# 07. Categorize daily busyness using a for-loop

# Verify the busyness of each day of the week

df_ords_prods_all['orders_day_of_week'].value_counts(dropna = False)

## Create "Busiest Day" column categorizing busiest and least busy days

# Create a "busiest day" function and populate with a busyness categorization

result = []

for value in df_ords_prods_all["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

df_ords_prods_all['busiest_day'] = result

df_ords_prods_all['busiest_day'].value_counts(dropna = False)

## Add "Busiest Days" column for top and bottom two days

### TASK 4.7 Step 2

# Create a "busiest day" function and populate with a busyness categorization of top two and bottom two busiest days

result2 = []

for value in df_ords_prods_all["orders_day_of_week"]:
  if value == 0 or value == 1:
    result2.append("Busiest day")
  elif value == 4 or value == 3:
    result2.append("Least busy")
  else:
    result2.append("Regularly busy")

result2

#Places results in new column

df_ords_prods_all['busiest_days'] = result2

# Check count of results

df_ords_prods_all['busiest_days'].value_counts(dropna = False)

### TASK 4.7 Step 3

**Expectations and Observations**

Expected:

- Busiest Days = 11,864,412
- Least Busy Days = 7,624,336

Results: As expected. Took awhile to figure out how to figure out the "OR" part vs creating new elif statements, but I got it! CELEBRATE!

# 08. Categorize hourly busyness using a for-loop

# Verify the busyness of each hour of the day

df_ords_prods_all['order_hour_of_day'].value_counts(dropna = False)

## TASK 4.7 Step 4

**Most and Least Orders Logic** I decided to look at the numbers and select a cut off for busiest and least busy based on significant changes from the order counts by hour. Based on the results above, it looks like the difference between hour 9 and hour 17 was significant, so I used that as the cutoff for busiest. Then looking at the least busy, the difference between hour 5 and 1 was significant, so I used that cutoff.

# Create a "busiest period of the day" function and populate with an hourly busyness categorization

result3 = []

for value in df_ords_prods_all["order_hour_of_day"]:
  if value == 10 or value == 11 or value == 14 or value == 15 or value == 13 or value == 12 or value == 16 or value == 9: 
    result3.append("Most orders")
  elif value == 5 or value == 2 or value == 4 or value == 3:
    result3.append("Least orders")
  else:
    result3.append("Average orders")

result3

#Places results in new column

df_ords_prods_all['busiest_period_of_day'] = result3

## TASK 4.7 Step 5

# Check count of results

df_ords_prods_all['busiest_period_of_day'].value_counts(dropna = False)

**Expectations and Observations**

Expected:

- Most Orders = 21,118,071
- Least Orders = 261,859

Results: As expected. Hooray!

# 09. Export updated df as pickle file

## TASK 4.7 Step 7

#Export updated df as pickle to Prepared Data folder

df_ords_prods_all.to_pickle(os.path.join(projpath, '02 Data', 'Prepared Data', 'ords_prods_derived.pkl'))
