==> 03 Scripts/6. Data Visualization - Part 1.ipynb <==
# 01. Notebook Setup

# Import libraries

import pandas as pd
import numpy as np
import os

# Create path to project folder

projpath = r'/Users/laineyodette/Documents/_proDocs/professional development/Learning/Learning - Data Analytics/CareerFoundry - Become a Data Analyst/Data Immersion Course/A4 - Python Fundamentals for Data Analysts/Submitted Tasks/2024-01 Instacart Basket Analysis'

# 02. File Import and DF Creation

# Import customers.csv file to create dataframe

df_cust = pd.read_csv(os.path.join(projpath, '02 Data', 'Original Data', 'customers.csv'), index_col = False)

# Import most recent ords_prods data file

ords_prods_agg = pd.read_pickle(os.path.join(projpath, '02 Data', 'Prepared Data', 'ords_prods_aggregation.pkl'))

# Check dimensions of ord_prods df - should have (32404859, 25) for shape

ords_prods_agg.shape

**Confirmed** ords_prods pulled in properly - all rows included.

# Check column headers of ords_prods df for later use in merge and to check data types

ords_prods_agg.info()

# 03. Quality / Consistency Checks + Data Wrangling

## Initial Quality Checks

# Check dimensions of new df

df_cust.shape

# Check output - view current column headers

df_cust.head(25)

# Check the values in the date_joined column - all seem to be the same? may not be relevant

df_cust['date_joined'].value_counts()

**Observation** 

date_joined column has different values. Keep column in case relevant for loyalty or time series evaluation.

# Check data types for each column

df_cust.info()

# Check for mixed data types

for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].map(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

# Verify data type of 'First Name' column

df_cust['First Name'].dtype

# Check output stats

df_cust.describe()

**Initial Evaluation and Planned Action** 

- *Columns not needed?* Plan to keep all columns - each may come in handy for analysis. Note that first/last name are not likely needed for analysis. But could come in handy in recommendations, if any would benefit from having the customer's name.

- *Columns to rename?* Plan to rename 'surnam' to 'last name', 'STATE' to 'state', 'n_dependants' to number_of_dependants', 'fam_status' to 'family_status'.

- *Data type changeds needed?* Plan to update 'user_id' to string. Need to figure out why 'First Name' is returning as a mixed data type. It appears to be a string, but comes back in the "weird" check as mixed. Also a note that the 'user_id' data type in the ords_prods_agg df should also therefore be string.

- *Missing values?* Plan to dive into 'First Name' column as the number of non-null values does not match the total number of rows so there may be some missing. Suspect that this may explain the mixed data type.

- *Any ususual values noticed from describe?* No, all values appear to be reasonable.

## Column Name Changes

# Rename the 'Surnam' column to be more intuitive

df_cust.rename(columns = {'Surnam' : 'Last Name'}, inplace = True)

# Rename the 'STATE' column to be more consistent to other headers

df_cust.rename(columns = {'STATE' : 'State'}, inplace = True)

# Rename the 'n_dependants' column to be more intuitive

df_cust.rename(columns = {'n_dependants' : 'Number_of_Dependants'}, inplace = True)

# Rename the 'fam_status' column to be more intuitive

df_cust.rename(columns = {'fam_status' : 'Family_Status'}, inplace = True)

# Re-check output - view updated column headers

df_cust.head()

**Confirmed** Column name changes complete.

## Data Type Changes

# Change the data type of the user_id column in customer df to string and override as new type

df_cust['user_id'] = df_cust['user_id'].astype('str')

# Check to ensure data type change has occurred

df_cust['user_id'].dtype

**Confirmed** Data type change complete.

**Note About 'First Name' Field** 

Suspect that there are missing values in this field, leading to the mixed data type. See next section for evaluation and any action.

# Change the data type of the user_id column in ords_prods_agg df to string and override as new type

ords_prods_agg['user_id'] = ords_prods_agg['user_id'].astype('str')

# Check to ensure data type change has occurred

ords_prods_agg['user_id'].dtype

**Confirmed** Data type change complete. Note - I did this already on this column back in an earlier task, so not sure why it didn't carry through until now.

## Missing Values Check

# Check for missing values in df

df_cust.isnull().sum()

**Observation and Response** 

As suspected, there are missing values in the 'First Name' field. It's about 5% of the data. 11,259 out of 206,209.

While this is a shame, the analysis does not require first names, so no action needed. Where this might become an issue is in recommendations. If there are any recommenation that would target specific customers, having a first name may be helpful. However, we still have the user_id to uniquely identify the customer, so it should be okay.

Also, a note that with the NaN values in this column, this explains the mixed data type returned from that check. Since this field will not be evaluated, no action needed to correct the data type. It is string with NaN values.

## Duplicate Values

# Look for full duplicates in customer df

df_cust_dups = df_cust[df_cust.duplicated()]

# Check for results in duplicate df

df_cust_dups

**Observations** No full duplicates found. No action required.

## Verify shape after all consistency and quality checks

# Verify shape after all consistency and quality checks

df_cust.shape

# 04. Merge Customer DF into Other Merged DF

- Note common column headers between two dfs (df_cust and ords_prods_agg)

- Use 'user_id' for merge key

# Remove current '_merge' column from ords_prods_agg df so that can check the new merge

ords_prods_agg = ords_prods_agg.drop(columns = ['_merge'])

# Verify column headers of ords_prods df to confirm _merge is removed

for col in ords_prods_agg.columns:
    print(col)

# Merge ords_prods_agg with df_cust using 'user_id' as key and include a merge column to indicate which df the row came from

df_merged = ords_prods_agg.merge(df_cust, on = 'user_id', indicator = True)

# Check count of merge flag to verify data came from both dfs

df_merged['_merge'].value_counts()

**Confirmed** Data came from both dfs. Same number of rows as ords_prods_agg base df.

# Check merged df column headers

df_merged.head()

# Check merged df shape

df_merged.shape

# 07. Export merged dataframe

# Export new merged dataframe as pickle file to Prepared Data folder

df_merged.to_pickle(os.path.join(projpath, '02 Data', 'Prepared Data', 'orders_products_all.pkl'))
